refactor: Move arrow key shortcuts into their own class.
